Organize device files into folder
