+++ src/components/device/Journal.tsx
