+++ src/containers/device/JournalContainer.ts
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
 
-import Journal from '../components/Journal'
+import Journal from '../../components/device/Journal'
 
 export default connect()(Journal)
